+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - #120: Added option to apply appropriate labels to Gmail messages after processing them.
 
+### Changed
+
+- #156: Circuit CID lookups are now case-insensitive
+
 ## v0.3.2 - 2021-10-26
 
 ### Fixed